+++ pages/wishes/category/[categoryId].js
@@ -1,7 +1,6 @@
 import React from 'react';
 import api from '../../../utils/api';
-//import ViewCategoryPage from '../../../src/components/category/pages/WishesViewCategoryPage';
-import ViewCategoryPage from '../../../src/components/category/pages/WishesViewCategoryPagea';
+import ViewCategoryPage from '../../../src/components/category/pages/WishesViewCategoryPage';
 
 import dynamic from 'next/dynamic';
 import SessionProvider from '../../../src/components/session/modules/SessionProvider';
@@ -19,7 +18,7 @@ export async function getServerSideProps({ params, query, req, res }) {
   return {
     props: {
       categoryDetails,
-      filterQuery: query.filter ? query.filter : null,
+      sortByQuery: query.sortBy ? query.sortBy : null,
       user,
     },
   };
@@ -33,14 +32,14 @@ const getCategoryDetails = async (categoryId) => {
   return rawCategory.data();
 };
 
-const ViewCategory = ({ categoryDetails, filterQuery, user }) => {
+const ViewCategory = ({ categoryDetails, sortByQuery, user }) => {
   if (Object.keys(categoryDetails).length === 0) {
     return <Error />;
   }
   return (
     <SessionProvider user={user}>
       <TopNavigationBar />
-      <ViewCategoryPage categoryDetails={categoryDetails} filterQuery={filterQuery} />
+      <ViewCategoryPage categoryDetails={categoryDetails} sortByQuery={sortByQuery} />
       <BottomNavigation />
     </SessionProvider>
   );
