+++ src/components/category/modules/WishesSortBy.js
@@ -1,6 +1,7 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { ChoiceGroup, Radio, Separator } from '@kiwicom/orbit-components/lib';
 import BlackText from '../../text/BlackText';
+import { useRouter } from 'next/router';
 
 /**
  * https://www.algolia.com/doc/api-reference/widgets/sort-by/react/#create-a-react-component
@@ -8,11 +9,29 @@ import BlackText from '../../text/BlackText';
  * @param {string} currentRefinement
  * @param {function} refine
  */
-const WishesSortBy = ({ items, currentRefinement, refine }) => {
+const WishesSortBy = ({ items, currentRefinement, refine, sortByIndex, category }) => {
+  const router = useRouter();
+
   const handleSelect = (event) => {
     const sortBy = event.target.value;
     refine(sortBy);
+    if (category) {
+      router.push(`/wishes/category/[categoryId]`, `/wishes/category/${category.id}?sortBy=${event.target.value}`, {
+        shallow: true,
+      });
+    } else {
+      // for view all wishes
+      router.push(`/wishes/category`, `/wishes/category?sortBy=${event.target.value}`, {
+        shallow: true,
+      });
+    }
   };
+
+  useEffect(() => {
+    if (sortByIndex) {
+      refine(sortByIndex)
+    }
+  }, [sortByIndex, refine])
 
   return (
     <>
